+++ prettyprint.sh
@@ -1,9 +1,11 @@
-#!/usr/bin/env bash
-
-# $@ - List of directories
+#!/bin/sh
 
 # Code beautifier with clang-format
 # Recursively parses the directories passed as command line arguments
+
+# Arguments:
+#
+# $@ - List of directories
 
 if test -z "$CLANG_FORMAT"; then
     echo "Please set the CLANG_FORMAT environment variable to point to the \
@@ -14,11 +16,11 @@ else
         echo "Error: $CLANG_FORMAT executable not found." >&2
         exit 1
     fi
-    echo "Code formatting with '$CLANG_FORMAT' the folders:"
+    echo "Code formatting with '$CLANG_FORMAT' the directories:"
 fi
 
-for folder in "$@"; do
-    echo "$folder"
-    find "$folder" \( -iname '*.cpp' -o -iname '*.c' -o -iname '*.h' \
+for directory in "$@"; do
+    echo "$directory"
+    find "$directory" \( -iname '*.cpp' -o -iname '*.c' -o -iname '*.h' \
         -o -iname '*.hpp' \) -exec "$CLANG_FORMAT" -style=file -i {} +
 done