+++ unit_tests/tests/classes/qengine.cpp
@@ -32,6 +32,9 @@ TEST(qpp_QEngineT_get_dit, AllTests) {}
 /// BEGIN std::vector<idx> QEngineT::get_dits() const
 TEST(qpp_QEngineT_get_dits, AllTests) {}
 
+/// BEGIN idx QEngineT::get_max_post_selection_reps() const
+TEST(qpp_QEngineT_get_max_post_selection_reps, AllTests) {}
+
 /// BEGIN bool QEngineT::get_measured_destructively() const
 TEST(qpp_QEngineT_get_measured_destructively, AllTests) {}
 
@@ -51,7 +54,8 @@ TEST(qpp_QEngineT_get_stats, AllTests) {}
 TEST(qpp_QEngineT_post_select_ok, AllTests) {}
 
 /// BEGIN QEngineT& QEngineT::reset(bool reset_stats = true,
-///       bool ensure_post_selection = false)
+///       bool ensure_post_selection = false,
+///       idx max_post_selection_reps = std::numeric_limits<idx>::max())
 TEST(qpp_QEngineT_reset, AllTests) {}
 
 /// BEGIN QEngineT& QEngineT::reset_stats()
@@ -62,6 +66,10 @@ TEST(qpp_QEngineT_set_dit, AllTests) {}
 
 /// BEGIN QEngineT& QEngineT::set_dits(std::vector<idx> dits)
 TEST(qpp_QEngineT_set_dits, AllTests) {}
+
+/// BEGIN QEngineT& QEngineT::set_max_post_selection_reps(
+///       idx max_post_selection_reps)
+TEST(qpp_QEngineT_set_max_post_selection_reps, AllTests) {}
 
 /// BEGIN QEngineT& QEngineT::set_state(const T& state) override
 TEST(qpp_QEngineT_set_state, AllTests) {}