+++ unit_tests/tests/classes/qnoisy_engine.cpp
@@ -22,7 +22,8 @@ TEST(qpp_QNoisyEngineT_execute, AllCircuitWithRepetitions) {}
 TEST(qpp_QNoisyEngineT_get_noise_results, AllTests) {}
 
 /// BEGIN QEngineT& QNoisyEngineT::reset(bool reset_stats = true,
-///       bool ensure_post_selection = false)
+///       bool ensure_post_selection = false, idx max_post_selection_reps =
+///       std::numeric_limits<idx>::max())
 TEST(qpp_QNoisyEngineT_reset, AllTests) {}
 
 /// BEGIN std::string QNoisyEngineT::traits_get_name() const override